pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wingman"
version = "0.1.0"
description = "Your personal AI wingman — powered by Kimi K2.5 + Ollama"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # Core
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "pydantic>=2.5.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",

    # LLM Providers
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "openai>=1.10.0",

    # Channels
    "python-telegram-bot>=21.0",
    "discord.py>=2.3.0",

    # Tools
    "playwright>=1.40.0",
    "duckduckgo-search>=4.0",
    "ddgs>=0.1.0",  # Added ddgs explicitly as per runtime warning

    # Memory & Logging
    "aiofiles>=23.0",
    "python-frontmatter>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0",
]

[project.scripts]
wingman = "src.cli.app:app"

[tool.hatch.build.targets.wheel]
packages = ["src"]
